fix: safeLong 소수점 파싱 오류 수정 (GP합산으로 TOT_SP에 소수 발생 시 0 출력 버그)
Long.parseLong → (long)Double.parseDouble 로 변경하여 소수점 포함 SP값 정상 처리

+++ src/main/java/my/prac/api/loa/controller/BossAttackController.java
@@ -5676,7 +5676,8 @@ public class BossAttackController {
 	}
 
 	private long safeLong(Object v) {
-		try { return v == null ? 0 : Long.parseLong(String.valueOf(v)); }
+		// Double 경유로 파싱해 "32661315485.5" 같은 소수점 값도 정상 처리
+		try { return v == null ? 0 : (long) Double.parseDouble(String.valueOf(v)); }
 		catch (Exception e) { return 0; }
 	}
 	private int safeInt(Object v) {
